requirements.txt
discord.py
aiohttp
numpy
opencv-python
paddleocr
paddlepaddle
Levenshtein
python-dotenv